cmu: Introduce enableable I2C0 clock
First attempts at an I2C implementation enabled their clock manualy from
the CMU register, but that proved to be inconvenient as there can't be a
delay (which requires a frozen clock) before the I2C clock is set up.

This is a first primitive splitting-out of the I2C clock from the
register to allow passing it into an I2C initialization.

The implementation is still unsound as it performs a read-modify-write
on a register that is not exclusively owned, see [1] for possible
mitigation:

[1]: https://github.com/japaric/svd2rust/issues/226