Use per-pin routing with pre-configuration on EFM32 devices
This allows phasing out the awkward whole-peripheral PWM construction
currently in place for those platforms.

This pulls in some run-time checking (as compared to the
fully-type-stated checks on the newer platforms), but prevents creating
a completely new parallel hiararchy of PreRoutedTimer<T> abstractions that
would then spawn a PreRoutedTimerChannel<T, C> which could be statically
routed into a RoutedTimerChannel -- with great verbosity and some
duplication.